folder structure, optional chaining, cookies refactor, readme
