+++ package.json
@@ -17,7 +17,6 @@
     "axios": "^1.6.8",
     "date-fns": "^3.6.0",
     "react": "^18.3.1",
-    "react-cookie": "^7.1.4",
     "react-dom": "^18.3.1",
     "react-router-dom": "^6.23.0",
     "react-scripts": "5.0.1",
