+++ src/App.tsx
@@ -1,8 +1,8 @@
 import { Route, Routes } from "react-router-dom";
-import { RouteGuard } from "./components/common/RouteGuard";
-import { Favorites } from "./pages/Favorites";
-import Home from "./pages/Home";
-import SignIn from "./pages/SignIn";
+import { RouteGuard } from "./components/RouteGuard/RouteGuard";
+import { Favorites } from "./pages/Favorites/Favorites";
+import Home from "./pages/Home/Home";
+import SignIn from "./pages/SignIn/SignIn";
 const App = () => {
   return (
     <Routes>
