+++ src/modules/auth/auth.constants.ts
@@ -1,0 +1,9 @@
+export const AUTH_COOKIE_DURATION = 60;
+
+export const AUTH_COOKIE_KEY = "auth_token";
+
+export const DUMMY_USERS = [
+  { username: "admin", password: "admin", auth_token: "admin" },
+  { username: "user", password: "user", auth_token: "user" },
+  { username: "x", password: "x", auth_token: "x" },
+];
