+++ src/modules/auth/auth.hooks.ts
@@ -1,0 +1,4 @@
+import { useCookies } from "../../utils/hooks/useCookies";
+import { AUTH_COOKIE_KEY } from "./auth.constants";
+
+export const useAuthCookie = () => useCookies({ key: AUTH_COOKIE_KEY });
