+++ src/modules/auth/auth.types.ts
