+++ src/modules/character/character.constants.ts
@@ -1,0 +1,11 @@
+export const SEARCHABLE_CHARACTER_FIELDS = [
+  "name",
+  "status",
+  "species",
+  "type",
+  "gender",
+];
+
+export const CHARACTER_BASE_URL = "https://rickandmortyapi.com/api";
+
+export const FAVORITE_CHARACTERS_IDS = "favorite_characters_ids";
