--- src/pages/Favorites.tsx
@@ -1,11 +1,0 @@
-import { FavoriteCharactersContainer } from "../components/character/FavoriteCharacterContainer";
-import { Header } from "../components/common/header/Header";
-
-export const Favorites = () => {
-  return (
-    <>
-      <Header />
-      <FavoriteCharactersContainer />
-    </>
-  );
-};
