+++ src/pages/Favorites/Favorites.tsx
@@ -1,0 +1,11 @@
+import { Header } from "../../components/Header/Header";
+import { FavoriteCharactersContainer } from "./components/FavoritesContainer/FavoritesContainer";
+
+export const Favorites = () => {
+  return (
+    <>
+      <Header />
+      <FavoriteCharactersContainer />
+    </>
+  );
+};
