--- src/pages/Home.tsx
@@ -1,13 +1,0 @@
-import { CharactersContainer } from "../components/character/CharacterContainer";
-import { Header } from "../components/common/header/Header";
-
-const Home = () => {
-  return (
-    <>
-      <Header />
-      <CharactersContainer />
-    </>
-  );
-};
-
-export default Home;
