+++ src/pages/Home/Home.tsx
@@ -1,0 +1,13 @@
+import { Header } from "../../components/Header/Header";
+import { CharactersContainer } from "./components/CharacterContainer/CharacterContainer";
+
+const Home = () => {
+  return (
+    <>
+      <Header />
+      <CharactersContainer />
+    </>
+  );
+};
+
+export default Home;
