--- src/utils/constants.ts
@@ -1,10 +1,0 @@
-export const AUTH_COOKIE_KEY = "auth_token";
-export const AUTH_COOKIE_DURATION = 60;
-export const FAVORITE_CHARACTERS_IDS = "favorite_characters_ids";
-export const SEARCHABLE_CHARACTER_FIELDS = [
-  "name",
-  "status",
-  "species",
-  "type",
-  "gender",
-];
